+++ libs/cpp_dbc/include/cpp_dbc/core/columnar/columnar_db_result_set.hpp
@@ -47,7 +47,7 @@ namespace cpp_dbc
     class ColumnarDBResultSet : public DBResultSet
     {
     public:
-        virtual ~ColumnarDBResultSet() = default;
+        ~ColumnarDBResultSet() override = default;
 
         // Row navigation
         /**