+++ libs/cpp_dbc/include/cpp_dbc/core/document/document_db_cursor.hpp
@@ -48,7 +48,7 @@ namespace cpp_dbc
     class DocumentDBCursor : public DBResultSet
     {
     public:
-        virtual ~DocumentDBCursor() = default;
+        ~DocumentDBCursor() override = default;
 
         // Navigation
         /**