+++ libs/cpp_dbc/include/cpp_dbc/core/relational/relational_db_connection.hpp
@@ -44,7 +44,7 @@ namespace cpp_dbc
     class RelationalDBConnection : public DBConnection
     {
     public:
-        virtual ~RelationalDBConnection() = default;
+        ~RelationalDBConnection() override = default;
 
         // SQL execution
         /**