+++ libs/cpp_dbc/include/cpp_dbc/core/relational/relational_db_result_set.hpp
@@ -46,7 +46,7 @@ namespace cpp_dbc
     class RelationalDBResultSet : public DBResultSet
     {
     public:
-        virtual ~RelationalDBResultSet() = default;
+        ~RelationalDBResultSet() override = default;
 
         // Row navigation
         /**